# services/openapi.yaml
openapi: 3.0.3
info:
  title: SME Cloud-Native Microservices API - Topic 10
  version: 1.0.0
  description: API contract for gateway routing, authorization enforcement, security testing, and API fuzzing.
servers:
  - url: https://localhost:8443
    description: Kong API Gateway HTTPS endpoint
  - url: http://localhost:8000
    description: Kong API Gateway HTTP endpoint for local testing
...
paths:
...
  /api/v1/admin/maintenance:
    post:
      tags:
        - Admin
      operationId: runMaintenanceFetch
      summary: Execute maintenance fetch operation
      description: This endpoint is used to demonstrate SSRF behavior in the vulnerable version and URL allowlist/blocklist enforcement in the fixed version.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceFetchRequest'
      responses:
        '200':
          description: Maintenance fetch executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceFetchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Request blocked by SSRF protection policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, expired, revoked, or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    MaintenanceFetchRequest:
      type: object
      required:
        - fetch_url
      properties:
        fetch_url:
          type: string
          format: uri
          example: "http://169.254.169.254/latest/meta-data/"
          description: URL used for SSRF simulation and protection testing.
    MaintenanceFetchResponse:
      type: object
      properties:
        fetched:
          type: boolean
          example: true
        target:
          type: string
          example: "http://169.254.169.254/latest/meta-data/"
        result:
          type: string
          example: "metadata response"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "forbidden"
        message:
          type: string
          example: "Access denied"
        correlation_id:
          type: string
          example: "req-123456"